+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.111
+
+- Scenario format converter
+
 # 0.110
 
 - Bug fix: destroyed vehicle should drop units/break infantry